feat: Implementar validación de autenticación y agregar middleware para proteger rutas
