+++ src/controllers/dashboardData/getCountFlashcardsByTheme.ts
@@ -9,24 +9,6 @@ import { VerifiedToken } from "../../models/interfaces/auth";
 export const getCountFlashcardsByTheme = async (req: Request, res: Response): Promise<void> => {
     try {
         const { userId } = req.params
-        const token = req.headers.authorization?.split(' ')[1]
-
-        if (!token) {
-            res.status(401).json({ message: 'Token de autorización no proporcionado' })
-            return
-        }
-
-        const verifiedToken = await verifyToken(token, {
-            jwtKey: process.env.CLERK_JWT_KEY,
-        })
-
-        const tokenData = verifiedToken as unknown as VerifiedToken
-
-        const userLevel = tokenData.pla.split(':')[1]
-        console.log('Token:', token)
-        console.log('Verified Token:', tokenData)
-        console.log('User Level:', userLevel)
-
         const data = await dashboardService.getCountFlashcardsByTheme(userId)
 
         if (!data.success) {
