+++ src/index.ts
@@ -6,6 +6,7 @@ import healthRouter from './routes/health'
 import helmet from 'helmet'
 import { limiter } from './services/rateLimiter'
 import dashboardRouter from './routes/dashboardRouter'
+import { validateAuth } from './middlewares/validateAuth'
 
 const PORT: number | string = process.env.PORT ?? 4444
 const app = express()
@@ -16,10 +17,10 @@ dotenv.config()
 app.use(express.json())
 app.use(helmet())
 app.disable('x-powered-by')
-app.use('/api/questions', limiter({ minuteDuration: 1, maxRequest: 5 }), appRouter)
-app.use('/api/user', limiter({ minuteDuration: 1, maxRequest: 20 }), userRouter)
-app.use('/api/health', limiter({ minuteDuration: 1, maxRequest: 3 }), healthRouter)
-app.use('/api/dashboard', limiter({ minuteDuration: 1, maxRequest: 20 }), dashboardRouter)
+app.use('/api/questions', validateAuth, limiter({ minuteDuration: 1, maxRequest: 5 }), appRouter)
+app.use('/api/user', validateAuth, limiter({ minuteDuration: 1, maxRequest: 20 }), userRouter)
+app.use('/api/health', validateAuth, limiter({ minuteDuration: 1, maxRequest: 3 }), healthRouter)
+app.use('/api/dashboard', validateAuth, limiter({ minuteDuration: 1, maxRequest: 20 }), dashboardRouter)
 
 app.get('/', (req: Request, res: Response): void => {
     res.json({ mensaje: "Bienvenido al punto de entrada de la API" })
